Реализован "первый этап" моделирования, в котором абитуриенты выбирают универы по баллам прошлых лет и кладут копии заявлений

+++ src/main/kotlin/modeling/dto/Statement.kt
@@ -1,6 +1,6 @@
 package modeling.dto
 
-class Statement(currentStudentId: Int, currentYGSNId: Int, currentScore: Int, currentState: State) {
+class Statement(currentStudentId: Int, currentYGSNId: Int, currentScore: Double, currentState: State) {
     var studentId = currentStudentId
     var ygsnId = currentYGSNId
     var score = currentScore
